\name{cartlm-package}
\alias{cartlm-package}
\alias{cartlm}
\docType{package}
\title{
\packageTitle{cartlm}
}
\description{
\packageDescription{cartlm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cartlm}
\packageIndices{cartlm}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{cartlm}

Maintainer: \packageMaintainer{cartlm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
