% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartLM.R
\name{cartLM}
\alias{cartLM}
\title{CART with Linear Model}
\usage{
cartLM(
  formula,
  data,
  max_depth = 3,
  min_samples_leaf = 20,
  min_samples_split = min_samples_leaf * 2,
  min_ig = 0.001,
  metric_func = NULL
)
}
\arguments{
\item{formula}{model formula}

\item{data}{data}

\item{max_depth}{max depth of tree}

\item{min_samples_leaf}{minimal samples in leaf}

\item{min_samples_split}{minimal samples in node to split}

\item{min_ig}{minimal information gain for split node}

\item{metric_func}{default for regression is variance dan for classification is gini index / gini impurity. other metric for classification is entrophy}
}
\value{
model, feature importance
}
\description{
CART with Linear Model
}
\examples{
cartLM(len ~ ., data = ToothGrowth)
}
