% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.DecisionTree}
\alias{predict.DecisionTree}
\title{Predict Method for Decision Tree model fits}
\usage{
\method{predict}{DecisionTree}(model, newdata, ...)
}
\arguments{
\item{model}{DecisionTree Model from package cartlm}

\item{newdata}{data.frame}

\item{...}{}
}
\value{
numeric vector of y prediction
}
\description{
Predict Method for Decision Tree model fits
}
\examples{
mod <- cartLM(len ~ ., data = ToothGrowth)
predict(mod, ToothGrowth, method = "mean")
}
